back-end/requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
docstring-to-markdown==0.11
filelock==3.12.0
flake8==6.0.0
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
fsspec==2023.5.0
greenlet==2.0.2
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
mccabe==0.7.0
mpmath==1.3.0
networkx==3.1
numpy==1.24.1
opencv-python-headless==4.5.5.64
packaging==23.0
Pillow==8.1.0
pycodestyle==2.10.0
pydocstyle==6.2.3
pyflakes==3.0.1
python-lsp-server==1.7.1
pytoolconfig==1.2.5
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
rope==1.7.0
SQLAlchemy==2.0.15
sympy==1.12
tokenizers==0.13.3
torch==1.13.1
tqdm==4.65.0
transformers==4.28.1
typing_extensions==4.6.2
urllib3==2.0.2
Werkzeug==2.3.4
whatthepatch==1.0.4
zipp==3.15.0
